🏗️ Move `src/components/Router/` folder to `src/routes/`
In the new feature-based architecture, each feature will have its own router. This one, in the `src/routes/` folder, contains the root router, so it must be at the root of the `src/` folder.
