+++ src/routes/NotFoundRoute.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { useDispatch } from 'react-redux'
-import { ErrorCode, setErrorPage } from '../../redux/reducers/errorPage'
+import { ErrorCode, setErrorPage } from '../redux/reducers/errorPage'
 
 export function NotFoundRoute() {
     const dispatch = useDispatch()
